server/requirements.txt
flask
dotenv
pyjwt
email_validator
flask_cors